Style fix in message module, added pagination, download and delete files from s3

+++ src/routers/message.py
@@ -1,22 +1,24 @@
 from datetime import datetime
 from fastapi import APIRouter, Depends, Query, UploadFile
+from fastapi.responses import FileResponse
 from typing import List
 from sqlalchemy.orm import load_only
 from collections import defaultdict
+from tempfile import TemporaryDirectory
 
-import traceback
-import logging
+import os.path as p
+
 
 from db import get_database, Session
 from auth import get_user, get_teacher
 from models.user import Account, AccountInfo
 from models.course import Course, CourseMessage, CourseFiles, CourseStatistics
 from schemas.enums import EnumAccountType
-from schemas.message import CreateMessage, GetMessage, UpdateMessage, GetAllCourseMessages, MessageAttachement, MessageCreatedData, GetAllCourseAttachments
+from schemas.message import (CreateMessage, GetMessage, UpdateMessage, GetAllCourseMessages, MessageAttachment,
+                             MessageCreatedData, GetAllCourseAttachments)
 
 import S3.s3 as s3
 import errors
-
 
 router = APIRouter()
 
@@ -33,7 +35,7 @@ async def create_course_message(course_id: int,
         raise errors.course_not_found()
     if user.account_type == EnumAccountType.teacher and course.owner != user.id:
         raise errors.access_denied()
-    
+
     message = CourseMessage(course_id=course_id,
                             account_id=user.id,
                             content=params.content,
@@ -41,7 +43,7 @@ async def create_course_message(course_id: int,
     db.add(message)
     db.commit()
 
-    return CourseCreatedData(message_id=message.id)
+    return MessageCreatedData(message_id=message.id)
 
 
 @router.put("/{course_id}/message/{message_id}", status_code=204,
@@ -59,7 +61,7 @@ async def update_course_message(course_id: int,
         raise errors.course_message_not_found()
     if user.account_type == EnumAccountType.teacher and message.course.owner != user.id:
         raise errors.access_denied()
-    
+
     if params.content is not None:
         message.content = params.content
     if params.title is not None:
@@ -79,18 +81,27 @@ async def get_course_message(course_id: int,
     message = db.query(CourseMessage).filter_by(id=message_id, course_id=course_id).first()
     if message is None:
         raise errors.course_message_not_found()
-    
-    attachments = db.query(CourseFiles).options(load_only(CourseFiles.id, CourseFiles.name)).filter_by(message_id=message_id).all()
+    # TODO Add check if user can get this message
+    attachments = (db.query(CourseFiles).
+                   options(load_only(CourseFiles.id,
+                                     CourseFiles.name)).
+                   filter_by(message_id=message_id).
+                   all())
     attachments_list = []
     for attachment in attachments:
-        attachments_list.append(MessageAttachement(
+        attachments_list.append(MessageAttachment(
             id=attachment.id,
             file_name=attachment.name,
         ))
- 
-    owner_info = db.query(AccountInfo).options(load_only(AccountInfo.name, AccountInfo.surname, AccountInfo.patronymic)).filter_by().first()
+
+    owner_info = (db.query(AccountInfo).
+                  options(load_only(AccountInfo.name,
+                                    AccountInfo.surname,
+                                    AccountInfo.patronymic)).
+                  filter_by(account_id=message.account_id).
+                  first())
     if owner_info.patronymic is None:
-        owner_name = owner_info.surname + " " + owner_info.name 
+        owner_name = owner_info.surname + " " + owner_info.name
     else:
         owner_name = owner_info.surname + " " + owner_info.name + " " + owner_info.patronymic
 
@@ -111,21 +122,22 @@ async def delete_course_message(course_id: int,
     """
     Delete course message 
     Note: all attachments will be also deleted
-    """    
+    """
     message = db.query(CourseMessage).filter_by(id=message_id, course_id=course_id).first()
     if message is None:
         raise errors.course_message_not_found()
     if user.account_type == EnumAccountType.teacher and message.course.owner != user.id:
         raise errors.access_denied()
-    
+
     attachments = db.query(CourseFiles).filter_by(message_id=message_id).all()
+    s3_session = s3.s3()
     for attachment in attachments:
-        # TODO add deletion of file from s3
+        if s3_session.has_file(attachment.s3_location):
+            s3_session.delete_file(attachment.s3_path)
         db.delete(attachment)
 
     db.delete(message)
     db.commit()
-    
 
 
 @router.get("/{course_id}/message/all", response_model=List[GetAllCourseMessages])
@@ -135,13 +147,23 @@ async def get_all_course_messages(course_id: int,
                                   user: Account = Depends(get_user),
                                   db: Session = Depends(get_database)):
     """Get limit number of messages with base info"""
-    messages = db.query(CourseMessage).filter_by(course_id=course_id).limit(limit).offset().all()  # TODO Finish pagination
+    messages = (db.query(CourseMessage).
+                filter_by(course_id=course_id).
+                order_by(CourseMessage.created_at.desc()).
+                limit(limit).
+                offset(page*limit).
+                all())
     result, owner_names = [], defaultdict()
     for message in messages:
         if message.account_id not in owner_names.keys():
-            owner_info = db.query(AccountInfo).options(load_only(AccountInfo.name, AccountInfo.surname, AccountInfo.patronymic)).filter_by().first()
+            owner_info = (db.query(AccountInfo).
+                          options(load_only(AccountInfo.name,
+                                            AccountInfo.surname,
+                                            AccountInfo.patronymic)).
+                          filter_by(account_id=message.account_id).
+                          first())
             if owner_info.patronymic is None:
-                owner_name = owner_info.surname + " " + owner_info.name 
+                owner_name = owner_info.surname + " " + owner_info.name
             else:
                 owner_name = owner_info.surname + " " + owner_info.name + " " + owner_info.patronymic
             owner_names[message.account_id] = owner_name
@@ -157,29 +179,29 @@ async def get_all_course_messages(course_id: int,
 @router.post("/{course_id}/message/{message_id}/attachment", status_code=201,
              responses=errors.with_errors(errors.course_message_not_found(),
                                           errors.access_denied()))
-async def create_message_attachemnt(course_id: int,
+async def create_message_attachment(course_id: int,
                                     message_id: int,
                                     file_title: str,
                                     file: UploadFile,
                                     user: Account = Depends(get_teacher),
                                     db: Session = Depends(get_database)):
-    """Download message attachment to s3"""
+    """Upload message attachment to s3"""
     message = db.query(CourseMessage).filter_by(id=message_id, course_id=course_id).first()
     if message is None:
         raise errors.course_message_not_found()
-    
+
     if user.account_type == EnumAccountType.teacher and message.account_id != user.id:
         raise errors.access_denied()
-    
+
     s3_session = s3.s3()
-    s3_session.upload_file(file.file.read(), file.filename)
+    s3_path = f"/attachments/{course_id}/message/{file.filename}"
+    s3_session.upload_file(file.file.read(), s3_path)
     attachment = CourseFiles(message_id=message_id,
                              name=file_title,
-                             s3_path=f"/attachments/{course_id}/message/{file.filename}",  #TODO check if upload link in bucket is correct
+                             s3_path=s3_path,
                              owner=user.id)
     db.add(attachment)
     db.commit()
-    
 
 
 @router.get("/{course_id}/message/{{message_id}}/attachment/all", response_model=List[GetAllCourseAttachments],
@@ -189,34 +211,34 @@ async def get_all_message_attachments(course_id: int,
                                       message_id: int,
                                       user: Account = Depends(get_user),
                                       db: Session = Depends(get_database)):
-    """Shows base info about all message attachemnts"""
+    """Shows base info about all message attachments"""
     # Check if message with attachment exists
     message = db.query(CourseMessage).filter_by(id=message_id, course_id=course_id).first()
     if message is None:
         raise errors.course_message_not_found()
-    # Check if has rights to see attachments
+    # Check if user has rights to see attachments
     if (db.query(CourseStatistics).
-        options(load_only(CourseStatistics.course_id, 
-                          CourseStatistics.account_id)).
-        filter_by(course_id=course_id, account_id=user.id).
-        first()) and user.account_type != EnumAccountType.admin and message.account_id != user.id:
+            options(load_only(CourseStatistics.course_id,
+                              CourseStatistics.account_id)).
+            filter_by(course_id=course_id, account_id=user.id).
+            first()) and user.account_type != EnumAccountType.admin and message.account_id != user.id:
         raise errors.access_denied()
-    
-    attachments = db.query(CourseFiles).options(load_only(CourseFiles.id, CourseFiles.name)).filter_by(message_id=message_id).all()
+
+    attachments = db.query(CourseFiles).options(load_only(CourseFiles.id, CourseFiles.name)).filter_by(
+        message_id=message_id).all()
     result = []
     for attachment in attachments:
         result.append(GetAllCourseAttachments(id=attachment.id, title=attachment.name))
 
     return result
-    
-    
 
 
 @router.delete("/{course_id}/message/{message_id}/attachment", status_code=204,
                responses=errors.with_errors(errors.course_message_not_found(),
                                             errors.access_denied(),
-                                            errors.attachment_not_found()))
-async def delete_message_attachemnt(course_id: int,
+                                            errors.attachment_not_found(),
+                                            errors.resource_not_found()))
+async def delete_message_attachment(course_id: int,
                                     message_id: int,
                                     attachment_id: int = Query(),
                                     user: Account = Depends(get_teacher),
@@ -225,53 +247,62 @@ async def delete_message_attachemnt(course_id: int,
     message = db.query(CourseMessage).filter_by(id=message_id, course_id=course_id).first()
     if message is None:
         raise errors.course_message_not_found()
-    
+
     if user.account_type == EnumAccountType.teacher and message.account_id != user.id:
         raise errors.access_denied()
-    
+
     attachment = (db.query(CourseFiles).
                   options(load_only(CourseFiles.id, CourseFiles.message_id, CourseFiles.s3_path)).
                   filter_by(id=attachment_id, message_id=message_id).
                   first())
     if attachment is None:
         raise errors.attachment_not_found()
-    
-    # TODO here should be delete from s3
+
+    # Delete file from s3
+    s3_session = s3.s3()
+    if not s3_session.has_file(attachment.s3_location):
+        raise errors.resource_not_found()
+    s3_session.delete_file(attachment.s3_path)
 
     db.delete(attachment)
     db.commit()
-    
 
 
-@router.get("/{course_id}/message/{message_id}/attachment", 
+@router.get("/{course_id}/message/{message_id}/attachment", response_class=FileResponse,
             responses=errors.with_errors(errors.access_denied(),
-                                         errors.course_message_not_found()))
+                                         errors.course_message_not_found(),
+                                         errors.resource_not_found()))
 async def download_attachment(course_id: int,
                               message_id: int,
                               attachment_id: int = Query(),
                               user: Account = Depends(get_user),
                               db: Session = Depends(get_database)):
-    """User for attachement dowload"""
+    """User for attachment download"""
     # Check if message with attachment exists
     message = db.query(CourseMessage).filter_by(id=message_id, course_id=course_id).first()
     if message is None:
         raise errors.course_message_not_found()
-    
-    # Check if has rights to download attachment
+
+    # Check if user has rights to download attachment
     if (db.query(CourseStatistics).
-        options(load_only(CourseStatistics.course_id, 
-                          CourseStatistics.account_id)).
-        filter_by(course_id=course_id, account_id=user.id).
-        first()) and user.account_type != EnumAccountType.admin and message.account_id != user.id:
+            options(load_only(CourseStatistics.course_id,
+                              CourseStatistics.account_id)).
+            filter_by(course_id=course_id, account_id=user.id).
+            first()) and user.account_type != EnumAccountType.admin and message.account_id != user.id:
         raise errors.access_denied()
-    
+
     attachment = (db.query(CourseFiles).
                   options(load_only(CourseFiles.id, CourseFiles.message_id, CourseFiles.s3_path)).
                   filter_by(id=attachment_id, message_id=message_id).
                   first())
     if attachment is None:
         raise errors.attachment_not_found()
-    
-    # TODO here should be generated s3 link and send via multipart
-    
-    
+
+    s3_session = s3.s3()
+    with TemporaryDirectory(prefix="download_") as tmp:
+        file_path = p.join(tmp, f"{attachment.name}.{attachment.s3_path.split(".")[-1]}")
+        with open(file_path, mode="w+b") as f:
+            if not s3_session.has_file(attachment.s3_location):
+                raise errors.resource_not_found()
+            s3_session.download_file(f, attachment.s3_path)
+            return FileResponse(file_path, media_type="multipart/form-data")
